Window dtor + Window::SetMousePosition
Yes, small commit since I spend more time doing useless things

- Window destructor fixes an other segfault when exiting the program
- Most other things done here are too irrelevant I think

+++ 3DRadSpace/Engine3DRadSpace/Point.cpp
@@ -64,5 +64,5 @@ Point operator*(float s, const Point& p)
 
 Point operator/(float s, const Point& p)
 {
-    return Point( s / p.X, s / p.Y);
+    return Point(int(s / p.X), int(s / p.Y));
 }
